project/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.2.0
psycopg2==2.9.4
pytils==0.4.1
sqlparse==0.4.3
